Split trait purity test into dedicated file with RequiresPhp 8.0
